backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
gunicorn==21.2.0
pydantic==2.4.2
sqlalchemy==2.0.32
python-dotenv==1.0.1
requests==2.31.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1 --only-binary=:all:
cryptography==42.0.7 --only-binary=:all:
openai==1.12.0openai==2.7.1
